+++ src/cluecode/copyrights_hint.py
@@ -159,4 +159,6 @@ tm
 univ
 upstream
 write
+cid:13
+cid:2
 '''.split()